# docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "MobileWallet backend",
    "description": "\nBackend **MobileWallet** — un rassemblement d'**agrégateurs** de paiement Mobile Money.\n\nChaque agrégateur (DigiKUNTZ aujourd'hui, d'autres demain) est un module exposant\ndeux capacités :\n\n- **navigateur IA** (`mode=browser`) : un agent Playwright + LLM pilote le checkout,\n  exécute le paiement et **déduit** le « curl replay » (requête `/charge` + `/verify`),\n  persisté comme *template* réutilisable.\n- **curl replay** (`mode=replay`) : rejoue le paiement sans navigateur via le template stocké.\n- **auto** (`mode=auto`, défaut) : tente le replay d'abord, puis bascule sur le navigateur\n  si le replay est non concluant (`fallback_browser`).\n\nToute tentative est auditée (table `transactions`) quand Supabase est configuré.\n",
    "version": "0.2"
  },
  "paths": {
...
    "/apps": {
      "post": {
        "tags": [
          "projects"
        ],
        "summary": "Créer une app (projet)",
        "description": "Crée une app pour VOTRE compte. Génère son webhook_secret HMAC (non exposé).",
        "operationId": "create_app_apps_post",
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "type": "string",
              "title": "Authorization"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AppCreateSelf"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "401": {
            "description": "JWT manquant ou invalide."
          },
          "503": {
            "description": "Supabase non configuré."
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "get": {
        "tags": [
          "projects"
        ],
        "summary": "Lister vos apps",
        "operationId": "list_apps_apps_get",
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "type": "string",
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "401": {
            "description": "JWT manquant ou invalide."
          },
          "503": {
            "description": "Supabase non configuré."
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/apps/{app_id}/keys": {
      "post": {
        "tags": [
          "projects"
        ],
        "summary": "Générer une clé API (clé en clair montrée une seule fois)",
        "operationId": "create_key_apps__app_id__keys_post",
        "parameters": [
          {
            "name": "app_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "App Id"
            }
          },
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "type": "string",
              "title": "Authorization"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ApiKeyCreate"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiKeyCreated"
                }
              }
            }
          },
          "401": {
            "description": "JWT manquant ou invalide."
          },
          "503": {
            "description": "Supabase non configuré."
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "get": {
        "tags": [
          "projects"
        ],
        "summary": "Lister les clés d'une app (sans la clé en clair)",
        "operationId": "list_keys_apps__app_id__keys_get",
        "parameters": [
          {
            "name": "app_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "App Id"
            }
          },
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "type": "string",
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "401": {
            "description": "JWT manquant ou invalide."
          },
          "503": {
            "description": "Supabase non configuré."
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/apps/{app_id}/keys/{key_id}/revoke": {
      "post": {
        "tags": [
          "projects"
        ],
        "summary": "Révoquer une clé API",
        "operationId": "revoke_key_apps__app_id__keys__key_id__revoke_post",
        "parameters": [
          {
            "name": "app_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "App Id"
            }
          },
          {
            "name": "key_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Key Id"
            }
          },
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "type": "string",
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "401": {
            "description": "JWT manquant ou invalide."
          },
          "503": {
            "description": "Supabase non configuré."
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/pay": {
      "post": {
        "tags": [
          "payments"
        ],
        "summary": "Exécuter un paiement",
        "description": "Exécute un paiement via l'agrégateur et le mode demandés.\n\nAuthentifié par clé API (`Authorization: Bearer <clé>`). La transaction est\nrattachée à l'app de la clé (isolation) ; `end_user_ref` est conservé tel quel.\n\n- **auto** (défaut) : replay d'abord ; bascule navigateur si non concluant\n  (selon `fallback_browser`).\n- **browser** : flux IA complet ; déduit et persiste le template curl.\n- **replay** : rejoue via le template stocké (409 si absent).",
        "operationId": "pay_pay_post",
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "type": "string",
              "title": "Authorization"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PayRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PayResponse"
                }
              }
            }
          },
          "404": {
            "description": "Agrégateur inconnu."
          },
          "400": {
            "description": "Mode invalide."
          },
          "422": {
            "description": "Réseau non supporté (renvoie la liste exacte attendue)."
          },
          "409": {
            "description": "Mode replay sans template (lancer mode=browser d'abord)."
          },
          "502": {
            "description": "Replay échoué et fallback navigateur désactivé."
          },
          "503": {
            "description": "Service de paiement amont temporairement indisponible (code=network_unavailable). Réessayer plus tard."
          }
        }
      }
    },
...
    "/balance": {
      "get": {
        "tags": [
          "transactions"
        ],
        "summary": "Solde de votre app",
        "description": "Solde logique de VOTRE app = somme de vos encaissements réussis − vos retraits.\n\nC'est le montant disponible pour un virement (`POST /payout`).",
        "operationId": "app_balance_balance_get",
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "type": "string",
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/admin/developers": {
      "post": {
        "tags": [
          "admin"
        ],
        "summary": "Créer un developer (compte intégrateur)",
        "operationId": "create_developer_admin_developers_post",
        "parameters": [
          {
            "name": "x-admin-key",
            "in": "header",
            "required": false,
            "schema": {
              "type": "string",
              "title": "X-Admin-Key"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DeveloperCreate"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "401": {
            "description": "X-Admin-Key invalide ou manquant."
          },
          "503": {
            "description": "Supabase non configuré."
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/admin/apps": {
      "post": {
        "tags": [
          "admin"
        ],
        "summary": "Créer une app pour un developer",
        "description": "Génère le `webhook_secret` HMAC de l'app (utilisé pour signer les webhooks).\nLe secret n'est pas renvoyé ; il sert côté serveur à la signature.",
        "operationId": "create_app_admin_apps_post",
        "parameters": [
          {
            "name": "x-admin-key",
            "in": "header",
            "required": false,
            "schema": {
              "type": "string",
              "title": "X-Admin-Key"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AppCreate"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "401": {
            "description": "X-Admin-Key invalide ou manquant."
          },
          "503": {
            "description": "Supabase non configuré."
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/admin/developers/{developer_id}/apps": {
      "get": {
        "tags": [
          "admin"
        ],
        "summary": "Lister les apps d'un developer",
        "operationId": "list_apps_admin_developers__developer_id__apps_get",
        "parameters": [
          {
            "name": "developer_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Developer Id"
            }
          },
          {
            "name": "x-admin-key",
            "in": "header",
            "required": false,
            "schema": {
              "type": "string",
              "title": "X-Admin-Key"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "401": {
            "description": "X-Admin-Key invalide ou manquant."
          },
          "503": {
            "description": "Supabase non configuré."
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/admin/apps/{app_id}/keys": {
      "post": {
        "tags": [
          "admin"
        ],
        "summary": "Générer une clé API pour une app (clé en clair montrée 1 fois)",
        "operationId": "create_key_admin_apps__app_id__keys_post",
        "parameters": [
          {
            "name": "app_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "App Id"
            }
          },
          {
            "name": "x-admin-key",
            "in": "header",
            "required": false,
            "schema": {
              "type": "string",
              "title": "X-Admin-Key"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ApiKeyCreate"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiKeyCreated"
                }
              }
            }
          },
          "401": {
            "description": "X-Admin-Key invalide ou manquant."
          },
          "503": {
            "description": "Supabase non configuré."
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "get": {
        "tags": [
          "admin"
        ],
        "summary": "Lister les clés d'une app (sans la clé en clair)",
        "operationId": "list_keys_admin_apps__app_id__keys_get",
        "parameters": [
          {
            "name": "app_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "App Id"
            }
          },
          {
            "name": "x-admin-key",
            "in": "header",
            "required": false,
            "schema": {
              "type": "string",
              "title": "X-Admin-Key"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "401": {
            "description": "X-Admin-Key invalide ou manquant."
          },
          "503": {
            "description": "Supabase non configuré."
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/admin/apps/{app_id}/keys/{key_id}/revoke": {
      "post": {
        "tags": [
          "admin"
        ],
        "summary": "Révoquer une clé API",
        "description": "Désactive la clé (effet immédiat : purge aussi le cache de résolution).",
        "operationId": "revoke_key_admin_apps__app_id__keys__key_id__revoke_post",
        "parameters": [
          {
            "name": "app_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "App Id"
            }
          },
          {
            "name": "key_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Key Id"
            }
          },
          {
            "name": "x-admin-key",
            "in": "header",
            "required": false,
            "schema": {
              "type": "string",
              "title": "X-Admin-Key"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "401": {
            "description": "X-Admin-Key invalide ou manquant."
          },
          "503": {
            "description": "Supabase non configuré."
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/admin/apps/{app_id}/balance": {
      "get": {
        "tags": [
          "admin"
        ],
        "summary": "Solde courant d'une app",
        "description": "Solde logique de l'app = somme des encaissements réussis − ses retraits.",
        "operationId": "admin_app_balance_admin_apps__app_id__balance_get",
        "parameters": [
          {
            "name": "app_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "App Id"
            }
          },
          {
            "name": "x-admin-key",
            "in": "header",
            "required": false,
            "schema": {
              "type": "string",
              "title": "X-Admin-Key"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "401": {
            "description": "X-Admin-Key invalide ou manquant."
          },
          "503": {
            "description": "Supabase non configuré."
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "AppCreateSelf": {
        "properties": {
          "name": {
            "type": "string",
            "title": "Name",
            "description": "Nom de l'app/projet.",
            "examples": [
              "Acme Checkout"
            ]
          }
        },
        "type": "object",
        "required": [
          "name"
        ],
        "title": "AppCreateSelf",
        "description": "Création d'une app par le dev (le developer_id vient du JWT, pas du corps)."
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
    },
    "securitySchemes": {
      "ApiKey": {
        "type": "http",
        "scheme": "bearer",
        "description": "Clé API d'app (paiements) : Authorization: Bearer <sk_…>."
      },
      "DevJWT": {
        "type": "http",
        "scheme": "bearer",
        "description": "JWT du compte developer (gestion des projets) : Authorization: Bearer <jwt>."
      },
      "AdminKey": {
        "type": "apiKey",
        "in": "header",
        "name": "X-Admin-Key",
        "description": "Clé admin pour la supervision des developers/apps/clés."
      }
    }
  },
...
}